blog/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.34.0
cron-descriptor==1.4.0
Django==4.1.7
django-celery-beat==2.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
pillow==10.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.7
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tblib==1.7.0
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
